Move mocks into its own package
